Enable Dockerfile to work in new location (build context).
Previous builds of the nginx Dockerfile were probably made in the old location inside moshimoji-backend. Now that the build context for the image will be moshimoji-nginx, paths inside the Dockerfile have been updatedto work relative to moshimoji-nginx, and a mount of the backend has been included so that static assets can be included in the build.